avcodec: Change get_pixels() to ptrdiff_t linesize
Found-by: ubitux

+++ libavcodec/avdct.h
@@ -64,7 +64,7 @@ typedef struct AVDCT {
 
     void (*get_pixels)(int16_t *block /* align 16 */,
                        const uint8_t *pixels /* align 8 */,
-                       int line_size);
+                       ptrdiff_t line_size);
 } AVDCT;
 
 /**
